+++ include/proofnetwork/jsonamqpclient.h
@@ -12,6 +12,9 @@ class PROOF_NETWORK_EXPORT JsonAmqpClient : public AbstractAmqpReceiver
     Q_DECLARE_PRIVATE(JsonAmqpClient)
 protected:
     explicit JsonAmqpClient(JsonAmqpClientPrivate &dd, QObject *parent = nullptr);
+    explicit JsonAmqpClient(QObject *parent = nullptr);
+    void setCustomJsonMessageHandler(
+        const std::function<void(const QJsonDocument &, const QString &, const QHash<QString, QVariant> &)> &&handler);
 };
 
 } // namespace Proof